% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspline.gen.R
\name{bspline.gen}
\alias{bspline.gen}
\title{Generating Multivariate Sieves with B-Splines}
\usage{
bspline.gen(
  data,
  test.data = NULL,
  n.basis = NULL,
  max.interaction = 1,
  spline.degree = 2
)
}
\arguments{
\item{data}{a data frame or a matrix containing the original variables.}

\item{test.data}{an optional data frame or matrix containing the same variables as the argument \code{data}. If \code{NULL}, no test dataset is considered.}

\item{n.basis}{a non-negative integer that specifies the number of B-spline basis for each variable. It corresponds to the argument \code{df} in command \code{\link[splines]{bs}}, and
is sum of the expected number of knots and spline degree plus 1 (because intercept is considered). Hence, the assigned \code{n.basis} should be greater than \code{spline.degree+1}, and if not,
it is automatically set to be \code{spline.degree+1+3} with 3 equal to the number of knots. If \code{n.basis=NULL}, it will be calculated according to the formulas established in this \href{https://github.com/ccfang2/Masters_Thesis}{thesis}.}

\item{max.interaction}{a positive integer that specifies the largest number of interacting variables in a single term of the resulting sieve. For example, \code{max.interaction=1} indicates no interaction among different variables. It shouldn't
be larger than the number of original variables.}

\item{spline.degree}{a positive integer that specifies the degree of the piecewise polynomial in B-splines. It corresponds to the argument \code{degree} in command \code{\link[splines]{bs}}. Default is 2 for quadratic splines.}
}
\value{
\code{bspline.gen} is based on the function \code{\link[splines]{bs}} in package \pkg{splines}. It generates multivariate sieves with B-splines, considering the interactions between basis of different variables. If \code{test.data} is supplied,
the output is a list of two data frames. One contains the resulting sieves for main dataset, and the other includes sieves for test dataset. Knots used to generate sieves for both datasets are the same, so the sieves of test dataset could be applied
for the evaluation of model estimated with sieves of main dataset. The data frame in the output of \code{bspline.gen} could be directly used as an input of the argument \code{x.sieve} in command \code{\link{dimada}}.
}
\description{
The \code{bspline.gen} command is used to generate multivariate sieves with B-splines. It considers the interactions among B-splines bases of different variables in a dataset.
In particular, if a user defines a test dataset which he hopes to use for his own model evaluation, the \code{bspline.gen} command also generates the corresponding sieves of
the test dataset with the same knots used for each variable of the main dataset. The command \code{\link[splines]{bs}} in package \pkg{splines} is applied to create B-Splines for each variable.
}
\note{
An internal function \code{create_index_matrix} in package \pkg{Sieve} is applied to help with the generation of multivariate sieves in function \code{bspline.gen}.
}
\examples{
# a data frame with 3 variables and 100 observations
set.seed(200)
df <- data.frame(a=runif(100), b=runif(100), c=runif(100))
# sieves with B-splines, considering full interactions of these 3 variables
sieve1 <- bspline.gen(data=df, test.data=NULL, n.basis=10, max.interaction=3)

# adding a test dataset with the same variables as the above main dataset
set.seed(200)
df.test <- data.frame(a=runif(100), b=runif(100), c=runif(100))
sieve2 <- bspline.gen(data=df, test.data=df.test, n.basis=10, max.interaction=3)
}
\seealso{
\link{poly.gen}; \link{cosine.gen}; \link{sine.gen}; \link{trig.gen}; \link{haar.gen}; \link{daubechies.gen}; \link{dimada}.
}
\author{
Chencheng Fang, Bonn Graduate School of Economics, University of Bonn. Email: \email{ccfang@uni-bonn.de}
}
