% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosine.gen.R
\name{cosine.gen}
\alias{cosine.gen}
\title{Generating Multivariate Sieves with Cosine Polynomials}
\usage{
cosine.gen(data, test.data = NULL, n.basis = NULL, max.interaction = 1)
}
\arguments{
\item{data}{a data frame or a matrix containing the original variables.}

\item{test.data}{an optional data frame or matrix containing the same variables as the argument \code{data}. If \code{NULL}, no test dataset is considered.}

\item{n.basis}{a non-negative integer that specifies the number of cosine polynomial bases for each variable. If \code{NULL}, it will be calculated according to the formulas established in this \href{https://github.com/ccfang2/Masters_Thesis}{thesis}.}

\item{max.interaction}{a positive integer that specifies the largest number of interacting variables in a single term of the resulting sieve. For example, \code{max.interaction=1} indicates no interaction among different variables. It shouldn't
be larger than the number of original variables.}
}
\value{
\code{cosine.gen} generates multivariate sieves with cosine polynomials, considering the interactions between basis of different variables. If \code{test.data} is supplied,
the output is a list of two data frames. One contains the resulting sieves for main dataset, and the other includes sieves for test dataset. \code{n.basis} used to generate sieves for both datasets are the same, so the sieves of test dataset could be applied
for the evaluation of model estimated with sieves of main dataset. The data frame in the output of \code{cosine.gen} could be directly used as an input of the argument \code{x.sieve} in command \code{\link{dimada}}.
}
\description{
The \code{cosine.gen} command is used to generate multivariate sieves with cosine polynomials. It considers the interactions among cosine polynomial bases of different variables in a dataset.
In particular, if a user defines a test dataset which he hopes to use for his own model evaluation, the \code{cosine.gen} command also generates the corresponding sieves of
the test dataset with the same number of basis used for each variable of the main dataset.
}
\note{
The function \code{create_index_matrix} in package \pkg{Sieve} is applied to help with the generation of multivariate sieves in function \code{cosine.gen}.
}
\examples{
# a data frame with 3 variables and 100 observations
set.seed(200)
df <- data.frame(a=runif(100), b=runif(100), c=runif(100))
# sieves with cosine polynomials, considering full interactions of these 3 variables
sieve1 <- cosine.gen(data=df, test.data=NULL, n.basis=10, max.interaction=3)

# adding a test dataset with the same variables as the above main dataset
set.seed(200)
df.test <- data.frame(a=runif(100), b=runif(100), c=runif(100))
sieve2 <- cosine.gen(data=df, test.data=df.test, n.basis=10, max.interaction=3)
}
\seealso{
\link{poly.gen}; \link{bspline.gen}; \link{sine.gen}; \link{trig.gen}; \link{haar.gen}; \link{daubechies.gen}; \link{dimada}.
}
\author{
Chencheng Fang, Bonn Graduate School of Economics, University of Bonn. Email: \email{ccfang@uni-bonn.de}
}
