% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimada.plot.R
\name{dimada.plot}
\alias{dimada.plot}
\title{Plotting the Output of Function \code{dimada}}
\usage{
dimada.plot(object, x = "nzeros")
}
\arguments{
\item{object}{an object of S3 class \code{"dimada"}, which is an output of the function \code{dimada}.}

\item{x}{a character string that specifies the x-axis. Available options are \code{"nzeros"} and \code{"lambdas"}. The former represents the numbers of non zero coefficients and the latter indicates lambdas
used in Lasso-type methods. Default is \code{"nzeros"}.}
}
\value{
An object of \code{ggplot2} plot is returned. Specifically, cross-validated approximation errors (i.e., mean squared errors) are plotted against the numbers of non zero coefficients or lambdas. A red star is
used to mark the selected lambda or the consequently chosen number of non-zero coefficients, and the value of corresponding mean squared error is denoted on the red star.
}
\description{
The \code{dimada.plot} command aims to quickly plot the output of function \code{dimada}. The plot depicts the change of cross-validated approximation errors against lambdas used in Lasso-type methods
or the consequent numbers of non zero coefficients. If more than one method is defined in the argument \code{methods} of function \code{dimada}, all subplots will be combined into a single plot.
}
\examples{
# plot the output 'dimada2' in the example of function 'dimada'
set.seed(200)
df <- data.frame(a=runif(100), b=runif(100), c=runif(100))
response2 <- sin(2*df[,1])+tan(df[,2])+log(df[,3])+rnorm(100,0.04)
dimada2 <- dimada(y=response2, x=df, basis="bspline", max.interaction=1,
                  methods=c("Lasso","adaLasso"))
(plot <- dimada.plot(dimada2,x="nzeros"))
}
\seealso{
\link{dimada}; \link{dimada.summary}.
}
\author{
Chencheng Fang, Bonn Graduate School of Economics, University of Bonn. Email: \email{ccfang@uni-bonn.de}
}
