% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimada.summary.R
\name{dimada.summary}
\alias{dimada.summary}
\title{Summarizing the Output of Function \code{dimada}}
\usage{
dimada.summary(object, verbose = TRUE)
}
\arguments{
\item{object}{an object of S3 class \code{"dimada"}, which is an output of the function \code{dimada}.}

\item{verbose}{a logical value that specifies whether \code{dimada.summary} should suppress the summary to be directly printed on console. If \code{FALSE}, no summary will be printed out directly but users can save it to
an object and later print it out. See example below. Default is \code{TRUE}.}
}
\value{
\code{dimada.summary} uses \code{cat()} to print out the summary of \code{"dimada"} object on console, if not suppressed. Also, \code{dimada.summary} saves the same output as a character vector invisibly, so users could use
\code{cat()} to print it out later at their demand, as shown in examples. The summary is composed of important information in the output of function \code{dimada}.
}
\description{
The \code{dimada.summary} command aims to quickly summarize the output of function \code{dimada}. The summary includes an overview of the sieve that is used in dimension adaptive estimation and the information
of selected model for each Lasso-type method and the consequent terms with non-zero coefficients. Moreover, as the selected terms are also used to fit an OLS model, the results of these post selection models are summarized as well.
}
\examples{
# summarize the output 'dimada2' in the example of function 'dimada'
# and save the summary to an object called 'summa'
set.seed(200)
df <- data.frame(a=runif(100), b=runif(100), c=runif(100))
response2 <- sin(2*df[,1])+tan(df[,2])+log(df[,3])+rnorm(100,0.04)
dimada2 <- dimada(y=response2, x=df, basis="bspline", max.interaction=1,
                  methods=c("Lasso","adaLasso"))
summary <- dimada.summary(dimada2,verbose=TRUE)

# print out the saved summary
cat(paste(summary,"\n"))
}
\seealso{
\link{dimada}; \link{dimada.plot}.
}
\author{
Chencheng Fang, Bonn Graduate School of Economics, University of Bonn. Email: \email{ccfang@uni-bonn.de}
}
