% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly.gen.R
\name{poly.gen}
\alias{poly.gen}
\title{Generating Multivariate Sieves with Power Series or Legendre Polynomials}
\usage{
poly.gen(
  data,
  test.data = NULL,
  n.basis = NULL,
  max.interaction = 1,
  legendre = TRUE
)
}
\arguments{
\item{data}{a data frame or a matrix containing the original variables.}

\item{test.data}{an optional data frame or matrix containing the same variables as the argument \code{data}. If \code{NULL}, no test dataset is considered.}

\item{n.basis}{a non-negative integer that specifies the number of basis for each variable. If \code{NULL}, it will be calculated according to the formulas established in this \href{https://github.com/ccfang2/Masters_Thesis}{thesis}.}

\item{max.interaction}{a positive integer that specifies the largest number of interacting variables in a single term of the resulting sieve. For example, \code{max.interaction=1} indicates no interaction among different variables. It shouldn't
be larger than the number of original variables.}

\item{legendre}{a logical value that indicates if power series or Legendre polynomials are used for construction of basis for each variable. If \code{TRUE}, Legendre polynomials are applied; otherwise, power series.}
}
\value{
\code{poly.gen} generates multivariate sieves with power series or Legendre polynomials, considering the interactions between basis of different variables. If \code{test.data} is supplied,
the output is a list of two data frames. One contains the resulting sieves for main dataset, and the other includes sieves for test dataset. \code{n.basis} used to generate sieves for both datasets are the same, so the sieves of test dataset could be applied
for the evaluation of model estimated with sieves of main dataset. The data frame in the output of \code{poly.gen} could be directly used as an input of the argument \code{x.sieve} in command \code{\link{dimada}}.
}
\description{
The \code{poly.gen} command is used to generate multivariate sieves with power series or Legendre polynomials bases. It considers the interactions among these bases of different variables in a dataset.
In particular, if a user defines a test dataset which he hopes to use for his own model evaluation, the \code{poly.gen} command also generates the corresponding sieves of
the test dataset with the same number of basis used for each variable of the main dataset. The command \code{\link[plde]{legendre_polynomial}} in package \pkg{plde} is applied to create Legendre polynomials.
}
\note{
The function \code{create_index_matrix} in package \pkg{Sieve} is applied to help with the generation of multivariate sieves in function \code{poly.gen}.
}
\examples{
# a data frame with 3 variables and 100 observations
set.seed(200)
df <- data.frame(a=runif(100), b=runif(100), c=runif(100))
# sieves with power series, considering full interactions of these 3 variables
sieve1 <- poly.gen(data=df, test.data=NULL, n.basis=10, max.interaction=3, legendre=FALSE)

# sieves with Legendre polynomials, considering no interactions of variables
# and only 5 basis for each variable
sieve2 <- poly.gen(data=df, test.data=NULL, n.basis=5, max.interaction=1, legendre=TRUE)
}
\seealso{
\link{bspline.gen}; \link{cosine.gen}; \link{sine.gen}; \link{trig.gen}; \link{haar.gen}; \link{daubechies.gen}; \link{dimada}.
}
\author{
Chencheng Fang, Bonn Graduate School of Economics, University of Bonn. Email: \email{ccfang@uni-bonn.de}
}
