% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trig.gen.R
\name{trig.gen}
\alias{trig.gen}
\title{Generating Multivariate Sieves with Trigonometric Polynomials}
\usage{
trig.gen(data, test.data = NULL, n.basis = NULL, max.interaction = 1)
}
\arguments{
\item{data}{a data frame or a matrix containing the original variables.}

\item{test.data}{an optional data frame or matrix containing the same variables as the argument \code{data}. If \code{NULL}, no test dataset is considered.}

\item{n.basis}{a non-negative integer that specifies the number of trigonometric polynomial basis for each variable. If \code{NULL}, it will be calculated according to the formulas established in this \href{https://github.com/ccfang2/Masters_Thesis}{thesis}.}

\item{max.interaction}{a positive integer that specifies the largest number of interacting variables in a single term of the resulting sieve. For example, \code{max.interaction=1} indicates no interaction among different variables. It shouldn't
be larger than the number of original variables.}
}
\value{
\code{trig.gen} generates multivariate sieves with trigonometric polynomials, considering the interactions between basis of different variables. If \code{test.data} is supplied,
the output is a list of two data frames. One contains the resulting sieves for main dataset, and the other includes sieves for test dataset. \code{n.basis} used to generate sieves for both datasets are the same, so the sieves of test dataset could be applied
for the evaluation of model estimated with sieves of main dataset. The data frame in the output of \code{trig.gen} could be directly used as an input of the argument \code{x.sieve} in command \code{\link{dimada}}.
}
\description{
The \code{trig.gen} command is used to generate multivariate sieves with trigonometric polynomials. It considers the interactions among trigonometric polynomial bases of different variables in a dataset.
In particular, if a user defines a test dataset which he hopes to use for his own model evaluation, the \code{trig.gen} command also generates the corresponding sieves of
the test dataset with the same number of basis used for each variable of the main dataset.
}
\note{
By definition, trigonometric polynomials are sum of cosine and sine polynomials. Hence, in the construction of \code{trig.gen} command, functions \code{cosine.gen} and \code{sine.gen} are called upon and the resulting sieves from them are then combined
to get the final sieve for \code{trig.gen}.
}
\examples{
# a data frame with 3 variables and 100 observations
set.seed(200)
df <- data.frame(a=runif(100), b=runif(100), c=runif(100))
# sieves with trigonometric polynomials, considering interactions of at most 2 variables
sieve1 <- trig.gen(data=df, test.data=NULL, n.basis=10, max.interaction=2)

# adding a test dataset with the same variables as the above main dataset
set.seed(200)
df.test <- data.frame(a=runif(100), b=runif(100), c=runif(100))
sieve2 <- trig.gen(data=df, test.data=df.test, n.basis=10, max.interaction=2)
}
\seealso{
\link{poly.gen}; \link{bspline.gen}; \link{cosine.gen}; \link{sine.gen}; \link{haar.gen}; \link{daubechies.gen}; \link{dimada}.
}
\author{
Chencheng Fang, Bonn Graduate School of Economics, University of Bonn. Email: \email{ccfang@uni-bonn.de}
}
